BCI_framework/requirements.txt
astroid==2.4.2
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.16.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
oscpy==0.5.0
pandas==1.0.3
Pillow==7.2.0
pylint==2.5.3
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.14.0
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
Werkzeug==1.0.1
wrapt==1.12.1
